# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/files/presign-upload": {
      "post": {
        "operationId": "presignUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "",
        "tags": [
          "files"
        ]
      }
    },
    "/files/{id}/complete": {
      "post": {
        "operationId": "completeUpload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "",
        "tags": [
          "files"
        ]
      }
    },
    "/files/{id}/download-url": {
      "get": {
        "operationId": "presignDownload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "",
        "tags": [
          "files"
        ]
      }
    },
...
  },
  "info": {
    "title": "test API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "PresignUploadDto": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string",
            "example": "report.pdf"
          },
          "contentType": {
            "type": "string",
            "example": "application/pdf"
          },
          "size": {
            "type": "number"
          }
        },
        "required": [
          "filename",
          "contentType"
        ]
      }
    }
  }
}